+++ Repository/DogmaRepository.php
@@ -38,13 +38,19 @@ class DogmaRepository extends \WordPress\EsiClient\Swagger {
     /**
      * Get a list of dogma attribute ids
      *
-     * @return array
+     * @return \WordPress\EsiClient\Model\Dogma\DogmaAttributes
      */
     public function dogmaAttributes() {
         $this->setEsiRoute($this->esiEndpoints['dogma_attributes']);
         $this->setEsiVersion('v1');
 
-        return \json_decode($this->callEsi());
+        $esiData = $this->callEsi();
+
+        if(!\is_null($esiData)) {
+            $returnValue = $this->mapArray(\json_encode(['attributes' => $esiData]), '\WordPress\EsiClient\Model\Dogma\DogmaAttributes');
+        }
+
+        return $returnValue;
     }
 
     /**